+++ ui/src/app/services/swagger-fetch-clients.ts
@@ -28,7 +28,6 @@ import {
   CohortBuilderApi,
   CohortReviewApi,
   CohortsApi,
-  ConceptsApi,
   ConceptSetsApi,
   Configuration as FetchConfiguration,
   DataSetApi,
@@ -88,7 +87,6 @@ export const cohortAnnotationDefinitionApi = bindCtor(CohortAnnotationDefinition
 export const cohortBuilderApi = bindCtor(CohortBuilderApi);
 export const cohortReviewApi = bindCtor(CohortReviewApi);
 export const cohortsApi = bindCtor(CohortsApi);
-export const conceptsApi = bindCtor(ConceptsApi);
 export const conceptSetsApi = bindCtor(ConceptSetsApi);
 export const dataSetApi = bindCtor(DataSetApi);
 export const featuredWorkspacesConfigApi = bindCtor(FeaturedWorkspacesConfigApi);
